fix[staging-provider]: Fix workspace credentials handling
